refactored to retrieve vote information from autotask webhook
